requirements.txt
discord.py==1.7.3
asyncpraw
discord-together==1.2.4
python-dotenv
